Make textwrap2 a submodule of hyphen
